\name{lancet1}
\docType{data}
\alias{lancet1}
\title{Data on the Mortality in Iraq from Roberts el al (2004)}
\description{
 A dataframe of the key cluster-level results. 
}
\usage{lancet1}
\format{dataframe}
\source{An .xls file downloaded from \url{http://timlambert.org} in
  February 2006 and included with this package.}
\references{
  "Mortality before and after the 2003 invasion of Iraq: cluster sample
  survey" by Les Roberts, Riyadh Lafta, Richard Garfield, Jamal
  Khudhairi, and Gilbert Burnham. The Lancet, 29 October 2004.  
}
\keyword{datasets}
